openstack: better support for BYON & custom API server VIP
This commit introduces a new field, `APIServerSubnetID`, in the OpenStack platform spec
to allow users to specify the subnet ID where the API server VIP should be created.

Key changes:
- Added `APIServerSubnetID` to the OpenStackPlatformSpec API definition.
- Run `make update` to update CRD manifests to include the new field.
- Extended the OpenStack cluster creation logic to accept `--openstack-api-server-subnet-id`.
- Modified the control-plane operator to annotate the API server service with the specified subnet ID.
- Implemented validation rules to ensure proper usage of networking options.

If `APIServerSubnetID` is not provided, the cloud provider will allocate a port on the default external network.

This enhancement will allow our customers to expose the API server VIP
on a tenant network that they pre-created, instead of relying on a
shared external network.